# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Green Ecolution Management System.",
        "title": "Green Space Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Green Ecolution",
            "url": "https://green-ecolution.de",
            "email": "info@green-ecolution.de"
        },
        "license": {
            "name": "AGPL",
            "url": "https://raw.githubusercontent.com/green-ecolution/green-ecolution/develop/LICENSE"
        },
        "version": "develop"
    },
    "paths": {
...
        "/v1/info": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get info about the app and the server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get info about the app",
                "operationId": "get-app-info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AppInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/plugin": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get a list of all registered plugins",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Get a list of all registered plugins",
                "operationId": "get-plugins-list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PluginListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a plugin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Register a plugin",
                "operationId": "register-plugin",
                "parameters": [
                    {
                        "description": "Plugin registration request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PluginRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClientToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/plugin/{plugin_slug}": {
            "get": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Get a plugin info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Get a plugin info",
                "operationId": "get-plugin-info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug of the plugin",
                        "name": "plugin_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Plugin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/plugin/{plugin_slug}/heartbeat": {
            "post": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Heartbeat for a plugin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Heartbeat for a plugin",
                "operationId": "plugin-heartbeat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the plugin specified by slug during registration",
                        "name": "plugin_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/plugin/{plugin_slug}/token/refresh": {
            "post": {
                "description": "Refresh plugin token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Refresh plugin token",
                "operationId": "refresh-plugin-token",
                "parameters": [
                    {
                        "description": "Plugin authentication",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PluginAuth"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Slug of the plugin",
                        "name": "plugin_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClientToken"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/plugin/{plugin_slug}/unregister": {
            "post": {
                "security": [
                    {
                        "Keycloak": []
                    }
                ],
                "description": "Unregister a plugin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Unregister a plugin",
                "operationId": "unregister-plugin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug of the plugin",
                        "name": "plugin_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AppInfo": {
            "type": "object",
            "required": [
                "buildTime",
                "git",
                "goVersion",
                "map",
                "server",
                "version"
            ],
            "properties": {
                "buildTime": {
                    "type": "string"
                },
                "git": {
                    "$ref": "#/definitions/GitInfo"
                },
                "goVersion": {
                    "type": "string"
                },
                "map": {
                    "$ref": "#/definitions/MapInfo"
                },
                "server": {
                    "$ref": "#/definitions/ServerInfo"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "ClientToken": {
            "type": "object",
            "required": [
                "access_token",
                "expires_in",
                "expiry",
                "id_token",
                "not_before_policy",
                "refresh_expires_in",
                "refresh_token",
                "scope",
                "session_state",
                "token_type"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "expiry": {
                    "type": "string"
                },
                "id_token": {
                    "type": "string"
                },
                "not_before_policy": {
                    "type": "integer"
                },
                "refresh_expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "session_state": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "GitInfo": {
            "type": "object",
            "required": [
                "branch",
                "commit",
                "repository"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "commit": {
                    "type": "string"
                },
                "repository": {
                    "type": "string"
                }
            }
        },
        "HTTPError": {
            "type": "object",
            "required": [
                "code",
                "error",
                "method",
                "path"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "MapInfo": {
            "type": "object",
            "required": [
                "bbox",
                "center"
            ],
            "properties": {
                "bbox": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "center": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "Plugin": {
            "type": "object",
            "required": [
                "description",
                "host_path",
                "name",
                "slug",
                "version"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "host_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "PluginAuth": {
            "type": "object",
            "required": [
                "client_id",
                "client_secret"
            ],
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                }
            }
        },
        "PluginListResponse": {
            "type": "object",
            "required": [
                "plugins"
            ],
            "properties": {
                "plugins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Plugin"
                    }
                }
            }
        },
        "PluginRegisterRequest": {
            "type": "object",
            "required": [
                "auth",
                "description",
                "name",
                "path",
                "slug",
                "version"
            ],
            "properties": {
                "auth": {
                    "$ref": "#/definitions/PluginAuth"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "ServerInfo": {
            "type": "object",
            "required": [
                "arch",
                "hostname",
                "interface",
                "ip",
                "os",
                "port",
                "uptime",
                "url"
            ],
            "properties": {
                "arch": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "interface": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "uptime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Keycloak": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth",
            "tokenUrl": "https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token"
        }
    }
}